third_party/blink/web_tests/external/wpt/largest-contentful-paint/idlharness-expected.txt

This is a testharness.js-based test.
[FAIL] LargestContentfulPaint interface: default toJSON operation on lcp
  assert_equals: expected "number" but got "string"
Harness: the test ran to completion.
